Move commonly used constants to utils in manager dashboard
